+++ core/routes/__init__.py
@@ -1,5 +1,6 @@
 from .callback import callback_namespace
 from .debug import debug_namespace
+from .influxdb2 import influxdb2_namespace
 from .library import library_namespace
 from .sites import sites_namespace
 from .trial_network import trial_network_namespace
@@ -8,6 +9,7 @@ from .user import user_namespace
 __all__ = [
     "callback_namespace",
     "debug_namespace",
+    "influxdb2_namespace",
     "library_namespace",
     "sites_namespace",
     "trial_network_namespace",