+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [unreleased] - 2025-09-XX
+
+### Added
+
+- New namespace `influxdb2` for handler InfluxDB requests.
+
+### Changed
 ## [v0.5.2] - 2025-05-16
 
 ### Fixed
@@ -305,6 +312,7 @@
 
 - Frontend implementation.
 
+[unreleased]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.2...HEAD
 [v0.5.2]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.1...v0.5.2
 [v0.5.1]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/6G-SANDBOX/TNLCM/compare/v0.4.5...v0.5.0